+++ CHANGES.md
@@ -136,8 +136,8 @@
     -   Fix wrong setting major-mode issue at following wiki link([GH-427][])
     -   Fix not consider `markdown-list-indent-width` issue([GH-405][])
     -   Fix URL open issue which contains end parentheses ([GH-408][])
+    -   Follow link even if it is in header([GH-430][])
 
-  [gh-349]: https://github.com/jrblevin/markdown-mode/issues/349]
   [gh-171]: https://github.com/jrblevin/markdown-mode/issues/171
   [gh-216]: https://github.com/jrblevin/markdown-mode/issues/216
   [gh-222]: https://github.com/jrblevin/markdown-mode/issues/222
@@ -186,6 +186,7 @@
   [gh-331]: https://github.com/jrblevin/markdown-mode/issues/331
   [gh-335]: https://github.com/jrblevin/markdown-mode/pull/335
   [gh-340]: https://github.com/jrblevin/markdown-mode/issues/340
+  [gh-349]: https://github.com/jrblevin/markdown-mode/issues/349
   [gh-350]: https://github.com/jrblevin/markdown-mode/pull/350
   [gh-369]: https://github.com/jrblevin/markdown-mode/pull/369
   [gh-378]: https://github.com/jrblevin/markdown-mode/pull/378
@@ -195,6 +196,7 @@
   [gh-410]: https://github.com/jrblevin/markdown-mode/issues/410
   [gh-421]: https://github.com/jrblevin/markdown-mode/issues/421
   [gh-427]: https://github.com/jrblevin/markdown-mode/issues/427
+  [gh-430]: https://github.com/jrblevin/markdown-mode/issues/430
   [gh-437]: https://github.com/jrblevin/markdown-mode/issues/437
   [gh-451]: https://github.com/jrblevin/markdown-mode/issues/451